+++ packages/twenty-server/src/modules/messaging/message-import-manager/drivers/exceptions/message-network.exception.ts
@@ -1,0 +1,7 @@
+export enum MessageNetworkExceptionCode {
+  ECONNRESET = 'ECONNRESET',
+  ENOTFOUND = 'ENOTFOUND',
+  ECONNABORTED = 'ECONNABORTED',
+  ETIMEDOUT = 'ETIMEDOUT',
+  ERR_NETWORK = 'ERR_NETWORK',
+}
